refactor(parser): extract expression parsers into parser/expressions/ (part 1/2)
Mechanical move of per-expression-form parsers out of the 12k-line
parser/mod.rs into focused submodules under parser/expressions/:

- primary.rs       (parse_primary_eval)
- pipe.rs          (parse_pipe_expr / parse_compose_expr / parse_coalesce_expr)
- for_expr.rs      (parse_for_expr + ForBinding/ForBodyResult + helpers)
- if_expr.rs       (parse_if_expr / parse_if_body* / IfBodyResult)
- validate_expr.rs (parse_validate_expr + CompareOp)
- string_literal.rs (parse_string_value/literal + unescape helpers)

ParseContext, AST types (ValidateExpr, etc.), is_static_value /
is_static_eval / evaluate_static_value, parse_resource_expr / read_resource_expr /
module_call / block_contents stay in mod.rs and move in part 2 (#2262).

Public API unchanged: ForBinding and CompareOp are pub-use'd back to
carina_core::parser. mod.rs drops from 12,614 to 11,145 lines.

Refs #2262 (part 1 of 2)
