+++ carina-core/src/parser/mod.rs
@@ -3,8 +3,20 @@
 //! Convert DSL to AST using pest
 
 mod config;
+mod expressions;
 
 pub use config::{DecryptorFn, ProviderContext, ValidatorFn};
+pub use expressions::for_expr::ForBinding;
+pub use expressions::validate_expr::CompareOp;
+
+use expressions::for_expr::{extract_for_iterable_name, parse_for_expr};
+use expressions::if_expr::parse_if_expr;
+use expressions::pipe::parse_coalesce_expr;
+use expressions::primary::parse_primary_eval;
+use expressions::string_literal::{
+    parse_string_literal, parse_string_value, unescape_single_quoted,
+};
+use expressions::validate_expr::parse_validate_expr;
 
 use indexmap::IndexMap;
 
@@ -290,16 +302,7 @@ pub struct ValidationBlock {
     pub error_message: Option<String>,
 }
 
-/// Comparison operator in a validate expression
-#[derive(Debug, Clone, PartialEq)]
-pub enum CompareOp {
-    Gte,
-    Lte,
-    Gt,
-    Lt,
-    Eq,
-    Ne,
-}
+// `CompareOp` lives in `expressions::validate_expr` (re-exported above).
 
 /// Validate expression AST node
 #[derive(Debug, Clone, PartialEq)]
@@ -1296,133 +1299,6 @@ fn parse_require_statement(pair: pest::iterators::Pair<Rule>) -> Result<RequireB
     })
 }
 
-/// Parse a validate expression (boolean expression with comparisons and logical operators)
-fn parse_validate_expr(pair: pest::iterators::Pair<Rule>) -> Result<ValidateExpr, ParseError> {
-    match pair.as_rule() {
-        Rule::validate_expr => {
-            let inner = first_inner(pair, "validate_or_expr", "validate_expr")?;
-            parse_validate_expr(inner)
-        }
-        Rule::validate_or_expr => {
-            let mut inner = pair.into_inner();
-            let first = next_pair(&mut inner, "validate_and_expr", "validate_or_expr")?;
-            let mut result = parse_validate_expr(first)?;
-            for next in inner {
-                let right = parse_validate_expr(next)?;
-                result = ValidateExpr::Or(Box::new(result), Box::new(right));
-            }
-            Ok(result)
-        }
-        Rule::validate_and_expr => {
-            let mut inner = pair.into_inner();
-            let first = next_pair(&mut inner, "validate_not_expr", "validate_and_expr")?;
-            let mut result = parse_validate_expr(first)?;
-            for next in inner {
-                let right = parse_validate_expr(next)?;
-                result = ValidateExpr::And(Box::new(result), Box::new(right));
-            }
-            Ok(result)
-        }
-        Rule::validate_not_expr => {
-            let mut inner = pair.into_inner();
-            let first = next_pair(&mut inner, "operand", "validate_not_expr")?;
-            if first.as_rule() == Rule::validate_not_expr {
-                // This is the "!" ~ validate_not_expr branch
-                let operand = parse_validate_expr(first)?;
-                Ok(ValidateExpr::Not(Box::new(operand)))
-            } else {
-                // This is the validate_comparison branch
-                parse_validate_expr(first)
-            }
-        }
-        Rule::validate_comparison => {
-            let mut inner = pair.into_inner();
-            let lhs_pair = next_pair(&mut inner, "validate_primary", "validate_comparison")?;
-            let lhs = parse_validate_expr(lhs_pair)?;
-            if let Some(op_pair) = inner.next() {
-                let op = match op_pair.as_str() {
-                    ">=" => CompareOp::Gte,
-                    "<=" => CompareOp::Lte,
-                    ">" => CompareOp::Gt,
-                    "<" => CompareOp::Lt,
-                    "==" => CompareOp::Eq,
-                    "!=" => CompareOp::Ne,
-                    other => {
-                        return Err(ParseError::InvalidExpression {
-                            line: 0,
-                            message: format!("Unknown comparison operator: {}", other),
-                        });
-                    }
-                };
-                let rhs_pair =
-                    next_pair(&mut inner, "validate_primary", "validate_comparison rhs")?;
-                let rhs = parse_validate_expr(rhs_pair)?;
-                Ok(ValidateExpr::Compare {
-                    lhs: Box::new(lhs),
-                    op,
-                    rhs: Box::new(rhs),
-                })
-            } else {
-                Ok(lhs)
-            }
-        }
-        Rule::validate_primary => {
-            let inner = first_inner(pair, "value", "validate_primary")?;
-            parse_validate_expr(inner)
-        }
-        Rule::validate_function_call => {
-            let mut inner = pair.into_inner();
-            let name = next_pair(&mut inner, "function name", "validate_function_call")?
-                .as_str()
-                .to_string();
-            let mut args = Vec::new();
-            for arg_pair in inner {
-                args.push(parse_validate_expr(arg_pair)?);
-            }
-            Ok(ValidateExpr::FunctionCall { name, args })
-        }
-        Rule::null_literal => Ok(ValidateExpr::Null),
-        Rule::boolean => Ok(ValidateExpr::Bool(pair.as_str() == "true")),
-        Rule::float => {
-            let f: f64 = pair
-                .as_str()
-                .parse()
-                .map_err(|e| ParseError::InvalidExpression {
-                    line: 0,
-                    message: format!("Invalid float: {}", e),
-                })?;
-            Ok(ValidateExpr::Float(f))
-        }
-        Rule::number => {
-            let n: i64 = pair
-                .as_str()
-                .parse()
-                .map_err(|e| ParseError::InvalidExpression {
-                    line: 0,
-                    message: format!("Invalid number: {}", e),
-                })?;
-            Ok(ValidateExpr::Int(n))
-        }
-        Rule::string => {
-            // Simple string parsing (no interpolation support in validate expressions)
-            let raw = pair.as_str();
-            // Strip surrounding quotes
-            let s = &raw[1..raw.len() - 1];
-            Ok(ValidateExpr::String(s.to_string()))
-        }
-        Rule::variable_ref => {
-            // Variable reference - just the identifier name
-            // For validate expressions, we only support simple variable names
-            let inner = first_inner(pair, "identifier", "variable_ref")?;
-            Ok(ValidateExpr::Var(inner.as_str().to_string()))
-        }
-        other => Err(ParseError::InvalidExpression {
-            line: 0,
-            message: format!("Unexpected rule in validate expression: {:?}", other),
-        }),
-    }
-}
-
 /// Parse attributes block
 fn parse_attributes_block(
     pair: pest::iterators::Pair<Rule>,
@@ -2059,461 +1935,6 @@ fn parse_primary_with_resource_or_module(
     }
 }
 
-/// Binding pattern for a for expression
-#[derive(Debug, Clone, PartialEq)]
-pub enum ForBinding {
-    /// Simple: `for x in ...`
-    Simple(String),
-    /// Indexed: `for (i, x) in ...`
-    Indexed(String, String),
-    /// Map: `for k, v in ...`
-    Map(String, String),
-}
-
-impl ForBinding {
-    /// Every binding name introduced by this pattern, in declaration order.
-    pub fn names(&self) -> Vec<&str> {
-        match self {
-            ForBinding::Simple(a) => vec![a.as_str()],
-            ForBinding::Indexed(a, b) | ForBinding::Map(a, b) => vec![a.as_str(), b.as_str()],
-        }
-    }
-}
-
-/// Whether `name` appears in `text` as a whole identifier (not a substring
-/// of a longer identifier). An identifier is bounded by anything that is
-/// not ASCII alphanumeric or `_`.
-fn identifier_appears_in(text: &str, name: &str) -> bool {
-    if name.is_empty() {
-        return false;
-    }
-    let bytes = text.as_bytes();
-    let name_bytes = name.as_bytes();
-    let is_id_char = |b: u8| b.is_ascii_alphanumeric() || b == b'_';
-
-    let mut i = 0;
-    while i + name_bytes.len() <= bytes.len() {
-        if &bytes[i..i + name_bytes.len()] == name_bytes {
-            let before_ok = i == 0 || !is_id_char(bytes[i - 1]);
-            let after_idx = i + name_bytes.len();
-            let after_ok = after_idx == bytes.len() || !is_id_char(bytes[after_idx]);
-            if before_ok && after_ok {
-                return true;
-            }
-        }
-        i += 1;
-    }
-    false
-}
-
-/// Whether `s` is shaped like a bare Carina identifier — the first byte is
-/// `A-Za-z_` and the rest are `A-Za-z0-9_`. Used to recover from the
-/// parser's collapse of unresolved identifiers into `Value::String(s)`
-/// when we need to decide whether to render an error as "identifier" vs
-/// "string literal". See #2101.
-fn is_bare_identifier(s: &str) -> bool {
-    let mut bytes = s.bytes();
-    match bytes.next() {
-        Some(b) if b.is_ascii_alphabetic() || b == b'_' => {}
-        _ => return false,
-    }
-    bytes.all(|b| b.is_ascii_alphanumeric() || b == b'_')
-}
-
-/// Result of parsing a for expression body: either a resource or a module call
-enum ForBodyResult {
-    Resource(Box<Resource>),
-    ModuleCall(ModuleCall),
-}
-
-/// Parse a for expression and expand it into individual resources and/or module calls.
-///
-/// `for x in list { resource_expr }` expands to resources with addresses like
-/// `binding[0]`, `binding[1]`, etc.
-///
-/// Extract a binding name from a for-expression's iterable.
-///
-/// For `for x in orgs.accounts { ... }`, returns `_accounts`.
-/// For non-variable iterables (lists, function calls), falls back to `_for{N}`.
-fn extract_for_iterable_name(pair: &pest::iterators::Pair<Rule>, counter: usize) -> String {
-    let fallback = format!("_for{}", counter);
-    let mut inner = pair.clone().into_inner();
-    // Skip for_binding, take for_iterable
-    let iterable_pair = inner.nth(1);
-    let Some(iterable) = iterable_pair else {
-        return fallback;
-    };
-    // Only extract name from variable_ref iterables
-    let first_child = iterable.into_inner().next();
-    let Some(child) = first_child else {
-        return fallback;
-    };
-    if child.as_rule() != Rule::variable_ref {
-        return fallback;
-    }
-    // Take the last segment of the dotted path (e.g., "accounts" from "orgs.accounts")
-    let text = child.as_str().trim();
-    let last_segment = text.rsplit('.').next().unwrap_or(text);
-    format!("_{}", last_segment)
-}
-
-/// `for k, v in map { resource_expr }` expands to resources with addresses like
-/// `binding["key1"]`, `binding["key2"]`, etc.
-///
-/// When the body is a module call, each iteration produces a module call with
-/// a binding name like `binding[0]` or `binding["key"]`.
-fn parse_for_expr(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &mut ParseContext,
-    binding_name: &str,
-) -> Result<(Vec<Resource>, Vec<ModuleCall>), ParseError> {
-    let for_line = pair.as_span().start_pos().line_col().0;
-    let mut inner = pair.into_inner();
-
-    // Parse the binding pattern
-    let binding_pair = next_pair(&mut inner, "for binding", "for expression")?;
-    let binding = parse_for_binding(binding_pair)?;
-
-    // Parse the iterable expression
-    let iterable_pair = next_pair(&mut inner, "iterable", "for expression")?;
-    let iterable = parse_for_iterable(iterable_pair, ctx)?;
-
-    // Parse the body (we'll re-parse it for each iteration)
-    let body_pair = next_pair(&mut inner, "body", "for expression")?;
-
-    // Warn on loop variables never referenced in the body. `_` is a discard
-    // marker and is skipped. The check is a scan over the body's raw text
-    // for the binding name as a whole identifier — good enough because the
-    // grammar already guarantees identifiers appear at token boundaries.
-    let body_text = body_pair.as_str();
-    for var in binding.names() {
-        if var == "_" || identifier_appears_in(body_text, var) {
-            continue;
-        }
-        ctx.warnings.push(ParseWarning {
-            file: None,
-            line: for_line,
-            message: format!(
-                "for-loop binding '{}' is unused. Rename to '_' to suppress this warning.",
-                var
-            ),
-        });
-    }
-
-    let mut resources = Vec::new();
-    let mut module_calls = Vec::new();
-
-    let collect = |result: ForBodyResult,
-                   resources: &mut Vec<Resource>,
-                   module_calls: &mut Vec<ModuleCall>| {
-        match result {
-            ForBodyResult::Resource(r) => resources.push(*r),
-            ForBodyResult::ModuleCall(c) => module_calls.push(c),
-        }
-    };
-
-    // Helper: only register non-discard bindings so `_` is never addressable.
-    let bind = |c: &mut ParseContext, name: &str, v: Value| {
-        if name != "_" {
-            c.set_variable(name.to_string(), v);
-        }
-    };
-
-    // Expand based on iterable type
-    match (&binding, &iterable) {
-        (ForBinding::Simple(var), Value::List(items)) => {
-            for (i, item) in items.iter().enumerate() {
-                let address = format!("{}[{}]", binding_name, i);
-                let mut iter_ctx = ctx.clone();
-                bind(&mut iter_ctx, var, item.clone());
-                let result = parse_for_body(body_pair.clone(), &iter_ctx, &address)?;
-                collect(result, &mut resources, &mut module_calls);
-            }
-        }
-        (ForBinding::Indexed(idx_var, val_var), Value::List(items)) => {
-            for (i, item) in items.iter().enumerate() {
-                let address = format!("{}[{}]", binding_name, i);
-                let mut iter_ctx = ctx.clone();
-                bind(&mut iter_ctx, idx_var, Value::Int(i as i64));
-                bind(&mut iter_ctx, val_var, item.clone());
-                let result = parse_for_body(body_pair.clone(), &iter_ctx, &address)?;
-                collect(result, &mut resources, &mut module_calls);
-            }
-        }
-        (ForBinding::Map(key_var, val_var), Value::Map(map)) => {
-            // Sort keys for deterministic output
-            let mut keys: Vec<&String> = map.keys().collect();
-            keys.sort();
-            for key in keys {
-                let val = &map[key];
-                let address = format!("{}[\"{}\"]", binding_name, key);
-                let mut iter_ctx = ctx.clone();
-                bind(&mut iter_ctx, key_var, Value::String(key.clone()));
-                bind(&mut iter_ctx, val_var, val.clone());
-                let result = parse_for_body(body_pair.clone(), &iter_ctx, &address)?;
-                collect(result, &mut resources, &mut module_calls);
-            }
-        }
-        // Unresolved reference — defer expansion to plan/apply when the
-        // upstream values are loaded. Field validity is checked statically
-        // by `upstream_exports::check_upstream_state_field_references`; for
-        // a valid field the deferral is an implementation detail the user
-        // doesn't need to hear about at validate time.
-        (_, Value::ResourceRef { path }) => {
-            // Build the for-expression header string
-            let header = match &binding {
-                ForBinding::Simple(var) => {
-                    format!("for {} in {}", var, path.to_dot_string())
-                }
-                ForBinding::Indexed(idx, val) => {
-                    format!("for ({}, {}) in {}", idx, val, path.to_dot_string())
-                }
-                ForBinding::Map(k, v) => {
-                    format!("for {}, {} in {}", k, v, path.to_dot_string())
-                }
-            };
-
-            // Try to parse the body once with placeholder values for the loop
-            // variable(s) to extract the resource type and attribute template.
-            let mut template_ctx = ctx.clone();
-            let placeholder = || Value::String(DEFERRED_UPSTREAM_PLACEHOLDER.to_string());
-            let bind = |c: &mut ParseContext, name: &str, v: Value| {
-                if name != "_" {
-                    c.set_variable(name.to_string(), v);
-                }
-            };
-            match &binding {
-                ForBinding::Simple(var) => {
-                    bind(&mut template_ctx, var, placeholder());
-                }
-                ForBinding::Indexed(idx, val) => {
-                    bind(
-                        &mut template_ctx,
-                        idx,
-                        Value::String(DEFERRED_UPSTREAM_INDEX_PLACEHOLDER.to_string()),
-                    );
-                    bind(&mut template_ctx, val, placeholder());
-                }
-                ForBinding::Map(k, v) => {
-                    bind(
-                        &mut template_ctx,
-                        k,
-                        Value::String(DEFERRED_UPSTREAM_KEY_PLACEHOLDER.to_string()),
-                    );
-                    bind(&mut template_ctx, v, placeholder());
-                }
-            }
-
-            let address = format!("{}[?]", binding_name);
-            if let Ok(ForBodyResult::Resource(resource)) =
-                parse_for_body(body_pair, &template_ctx, &address)
-            {
-                let attrs: Vec<(String, Value)> = resource
-                    .attributes
-                    .iter()
-                    .filter(|(k, _)| !k.starts_with('_'))
-                    .map(|(k, expr)| (k.clone(), expr.0.clone()))
-                    .collect();
-                ctx.deferred_for_expressions.push(DeferredForExpression {
-                    file: None,
-                    line: for_line,
-                    header,
-                    resource_type: if resource.id.provider.is_empty() {
-                        resource.id.resource_type.clone()
-                    } else {
-                        format!("{}.{}", resource.id.provider, resource.id.resource_type)
-                    },
-                    attributes: attrs,
-                    binding_name: binding_name.to_string(),
-                    iterable_binding: path.binding().to_string(),
-                    iterable_attr: path.attribute().to_string(),
-                    binding: binding.clone(),
-                    template_resource: *resource,
-                });
-            }
-            // Return empty — the for body produces zero concrete resources
-        }
-        _ => {
-            // Special case: the parser collapses bare unresolved identifiers
-            // (e.g. `for _ in org { ... }`) into `Value::String("org")` — the
-            // same slot a quoted literal uses. Reporting those as
-            // `iterable is string "org"` is misleading: the user wrote an
-            // identifier, not a literal, and the likely fault is a typo for
-            // a known binding. Record them as deferred for-expressions so
-            // `check_identifier_scope` (which runs on the merged
-            // directory-wide ParsedFile, so cross-file upstream_state /
-            // module bindings are visible) can emit a proper
-            // UndefinedIdentifier with the did-you-mean machinery from
-            // #2038 / #2100. See #2101 / #2138.
-            if let Value::String(s) = &iterable
-                && is_bare_identifier(s)
-            {
-                let header = match &binding {
-                    ForBinding::Simple(var) => format!("for {} in {}", var, s),
-                    ForBinding::Indexed(idx, val) => format!("for ({}, {}) in {}", idx, val, s),
-                    ForBinding::Map(k, v) => format!("for {}, {} in {}", k, v, s),
-                };
-                let mut template_ctx = ctx.clone();
-                let placeholder = || Value::String(DEFERRED_UPSTREAM_PLACEHOLDER.to_string());
-                let bind = |c: &mut ParseContext, name: &str, v: Value| {
-                    if name != "_" {
-                        c.set_variable(name.to_string(), v);
-                    }
-                };
-                match &binding {
-                    ForBinding::Simple(var) => {
-                        bind(&mut template_ctx, var, placeholder());
-                    }
-                    ForBinding::Indexed(idx, val) => {
-                        bind(
-                            &mut template_ctx,
-                            idx,
-                            Value::String(DEFERRED_UPSTREAM_INDEX_PLACEHOLDER.to_string()),
-                        );
-                        bind(&mut template_ctx, val, placeholder());
-                    }
-                    ForBinding::Map(k, v) => {
-                        bind(
-                            &mut template_ctx,
-                            k,
-                            Value::String(DEFERRED_UPSTREAM_KEY_PLACEHOLDER.to_string()),
-                        );
-                        bind(&mut template_ctx, v, placeholder());
-                    }
-                }
-                let address = format!("{}[?]", binding_name);
-                if let Ok(ForBodyResult::Resource(resource)) =
-                    parse_for_body(body_pair, &template_ctx, &address)
-                {
-                    let attrs: Vec<(String, Value)> = resource
-                        .attributes
-                        .iter()
-                        .filter(|(k, _)| !k.starts_with('_'))
-                        .map(|(k, expr)| (k.clone(), expr.0.clone()))
-                        .collect();
-                    ctx.deferred_for_expressions.push(DeferredForExpression {
-                        file: None,
-                        line: for_line,
-                        header,
-                        resource_type: if resource.id.provider.is_empty() {
-                            resource.id.resource_type.clone()
-                        } else {
-                            format!("{}.{}", resource.id.provider, resource.id.resource_type)
-                        },
-                        attributes: attrs,
-                        binding_name: binding_name.to_string(),
-                        iterable_binding: s.clone(),
-                        iterable_attr: String::new(),
-                        binding: binding.clone(),
-                        template_resource: *resource,
-                    });
-                }
-                return Ok((resources, module_calls));
-            }
-            let iterable_type = match &iterable {
-                Value::String(s) => {
-                    format!("string \"{}\"", if s.len() > 50 { &s[..50] } else { s })
-                }
-                Value::Int(i) => format!("int {}", i),
-                Value::Float(f) => format!("float {}", f),
-                Value::Bool(b) => format!("bool {}", b),
-                Value::ResourceRef { path } => {
-                    format!("unresolved reference {}", path.to_dot_string())
-                }
-                Value::List(_) => "list".to_string(),
-                Value::Map(_) => "map".to_string(),
-                other => format!("{:?}", other),
-            };
-            let binding_type = match &binding {
-                ForBinding::Simple(var) => format!("`for {} in ...`", var),
-                ForBinding::Indexed(idx, val) => format!("`for {}, {} in ...`", idx, val),
-                ForBinding::Map(k, v) => format!("`for {}, {} in ...`", k, v),
-            };
-            let expected = match &binding {
-                ForBinding::Simple(_) | ForBinding::Indexed(_, _) => "list",
-                ForBinding::Map(_, _) => "map",
-            };
-            return Err(ParseError::InvalidExpression {
-                line: for_line,
-                message: format!(
-                    "{} — iterable is {} (expected {})",
-                    binding_type, iterable_type, expected
-                ),
-            });
-        }
-    }
-
-    Ok((resources, module_calls))
-}
-
-/// Parse a for binding pattern.
-///
-/// Each position accepts either an `identifier` or a `discard_pattern`
-/// (`_`). The text of the matched pair — either the identifier name or
-/// the literal `_` — is stored as the binding name. A `_` marker is not
-/// added to the parse-time scope and is exempt from the unused-binding
-/// warning; downstream code checks for `name == "_"` to enforce that.
-fn parse_for_binding(pair: pest::iterators::Pair<Rule>) -> Result<ForBinding, ParseError> {
-    let inner = first_inner(pair, "binding pattern", "for binding")?;
-    match inner.as_rule() {
-        Rule::for_simple_binding => {
-            let name = first_inner(inner, "identifier", "simple binding")?
-                .as_str()
-                .to_string();
-            Ok(ForBinding::Simple(name))
-        }
-        Rule::for_indexed_binding => {
-            let mut parts = inner.into_inner();
-            let idx = next_pair(&mut parts, "index variable", "indexed binding")?
-                .as_str()
-                .to_string();
-            let val = next_pair(&mut parts, "value variable", "indexed binding")?
-                .as_str()
-                .to_string();
-            Ok(ForBinding::Indexed(idx, val))
-        }
-        Rule::for_map_binding => {
-            let mut parts = inner.into_inner();
-            let key = next_pair(&mut parts, "key variable", "map binding")?
-                .as_str()
-                .to_string();
-            let val = next_pair(&mut parts, "value variable", "map binding")?
-                .as_str()
-                .to_string();
-            Ok(ForBinding::Map(key, val))
-        }
-        _ => Err(ParseError::InternalError {
-            expected: "for binding pattern".to_string(),
-            context: "for expression".to_string(),
-        }),
-    }
-}
-
-/// Parse the iterable part of a for expression
-///
-/// When the iterable is a function call with all statically-known arguments,
-/// the function is eagerly evaluated at parse time. If any argument depends on
-/// a runtime value (e.g. ResourceRef), a clear error is returned.
-fn parse_for_iterable(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-) -> Result<Value, ParseError> {
-    // for_iterable contains function_call | list | variable_ref | "(" expression ")"
-    let inner = first_inner(pair, "iterable expression", "for iterable")?;
-    let eval = parse_primary_eval(inner, ctx)?;
-    let value = eval
-        .into_value()
-        .map_err(|leak| ParseError::InvalidExpression {
-            line: 0,
-            message: format!(
-                "for iterable evaluates to a closure '{}' (still needs {} arg(s)); \
-             closures cannot be iterated",
-                leak.name, leak.remaining_arity
-            ),
-        })?;
-    evaluate_static_value(value, ctx.config)
-}
-
 /// Check whether a Value is fully static (no runtime dependencies).
 fn is_static_value(value: &Value) -> bool {
     match value {
@@ -2581,285 +2002,6 @@ fn evaluate_static_value(value: Value, config: &ProviderContext) -> Result<Value
         }
         other => Ok(other),
     }
-}
-
-/// Parse the body of a for expression and produce a single resource or module call
-fn parse_for_body(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-    address: &str,
-) -> Result<ForBodyResult, ParseError> {
-    let mut local_ctx = ctx.clone();
-
-    for inner in pair.into_inner() {
-        match inner.as_rule() {
-            Rule::for_local_binding => {
-                let mut binding_inner = inner.into_inner();
-                let name = next_pair(&mut binding_inner, "binding name", "for local binding")?
-                    .as_str()
-                    .to_string();
-                let value = parse_expression(
-                    next_pair(&mut binding_inner, "binding value", "for local binding")?,
-                    &local_ctx,
-                )?;
-                local_ctx.set_variable(name, value);
-            }
-            Rule::resource_expr => {
-                let resource = parse_resource_expr(inner, &local_ctx, address)?;
-                return Ok(ForBodyResult::Resource(Box::new(resource)));
-            }
-            Rule::read_resource_expr => {
-                let resource = parse_read_resource_expr(inner, &local_ctx, address)?;
-                return Ok(ForBodyResult::Resource(Box::new(resource)));
-            }
-            Rule::module_call => {
-                let mut call = parse_module_call(inner, &local_ctx)?;
-                call.binding_name = Some(address.to_string());
-                return Ok(ForBodyResult::ModuleCall(call));
-            }
-            _ => {}
-        }
-    }
-
-    Err(ParseError::InternalError {
-        expected: "resource expression or module call".to_string(),
-        context: "for body".to_string(),
-    })
-}
-
-/// Result of parsing an if expression body: a resource, a module call, or a value
-enum IfBodyResult {
-    Resource(Box<Resource>),
-    ModuleCall(ModuleCall),
-    Value(Value),
-}
-
-/// Parse an if expression and conditionally include resources/module calls/values.
-///
-/// `if condition { body }` includes the body when condition is true.
-/// `if condition { body } else { body }` selects one branch.
-///
-/// The condition must evaluate to a static Bool value at parse time.
-fn parse_if_expr(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &mut ParseContext,
-    binding_name: &str,
-) -> Result<LetBindingRhs, ParseError> {
-    let mut inner = pair.into_inner();
-
-    // Parse the condition expression
-    let condition_pair = next_pair(&mut inner, "condition", "if expression")?;
-    let condition_value = parse_expression(condition_pair, ctx)?;
-
-    // Ensure the condition is statically evaluable
-    if !is_static_value(&condition_value) {
-        return Err(ParseError::InvalidExpression {
-            line: 0,
-            message: "if condition depends on a runtime value; \
-                      condition must be statically known at parse time"
-                .to_string(),
-        });
-    }
-
-    let condition_value = evaluate_static_value(condition_value, ctx.config)?;
-
-    // Condition must be a Bool
-    let condition = match &condition_value {
-        Value::Bool(b) => *b,
-        other => {
-            return Err(ParseError::InvalidExpression {
-                line: 0,
-                message: format!("if condition must be a Bool value, got: {:?}", other),
-            });
-        }
-    };
-
-    // Parse the if body
-    let if_body_pair = next_pair(&mut inner, "if body", "if expression")?;
-
-    // Check for else clause
-    let else_body_pair = inner.next();
-
-    if condition {
-        // Use the if branch
-        parse_if_body_to_rhs(if_body_pair, ctx, binding_name)
-    } else if let Some(else_pair) = else_body_pair {
-        // Use the else branch
-        let else_body = first_inner(else_pair, "else body", "else clause")?;
-        parse_if_body_to_rhs(else_body, ctx, binding_name)
-    } else {
-        // No else clause and condition is false: produce nothing
-        let ref_value = Value::String(format!("${{if:{}}}", binding_name));
-        Ok((EvalValue::from_value(ref_value), vec![], vec![], None))
-    }
-}
-
-/// Parse an if/else body and convert the result to a LetBindingRhs
-fn parse_if_body_to_rhs(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-    binding_name: &str,
-) -> Result<LetBindingRhs, ParseError> {
-    let result = parse_if_body(pair, ctx, binding_name)?;
-    match result {
-        IfBodyResult::Resource(r) => {
-            let ref_value = Value::String(format!("${{{}}}", binding_name));
-            Ok((EvalValue::from_value(ref_value), vec![*r], vec![], None))
-        }
-        IfBodyResult::ModuleCall(c) => {
-            let value = Value::String(format!("${{module:{}}}", c.module_name));
-            Ok((EvalValue::from_value(value), vec![], vec![c], None))
-        }
-        IfBodyResult::Value(v) => Ok((EvalValue::from_value(v), vec![], vec![], None)),
-    }
-}
-
-/// Parse the body of an if expression and produce a resource, module call, or value
-fn parse_if_body(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-    binding_name: &str,
-) -> Result<IfBodyResult, ParseError> {
-    let mut local_ctx = ctx.clone();
-
-    for inner in pair.into_inner() {
-        match inner.as_rule() {
-            Rule::if_local_binding => {
-                let mut binding_inner = inner.into_inner();
-                let name = next_pair(&mut binding_inner, "binding name", "if local binding")?
-                    .as_str()
-                    .to_string();
-                let value = parse_expression(
-                    next_pair(&mut binding_inner, "binding value", "if local binding")?,
-                    &local_ctx,
-                )?;
-                local_ctx.set_variable(name, value);
-            }
-            Rule::resource_expr => {
-                let resource = parse_resource_expr(inner, &local_ctx, binding_name)?;
-                return Ok(IfBodyResult::Resource(Box::new(resource)));
-            }
-            Rule::read_resource_expr => {
-                let resource = parse_read_resource_expr(inner, &local_ctx, binding_name)?;
-                return Ok(IfBodyResult::Resource(Box::new(resource)));
-            }
-            Rule::module_call => {
-                let mut call = parse_module_call(inner, &local_ctx)?;
-                call.binding_name = Some(binding_name.to_string());
-                return Ok(IfBodyResult::ModuleCall(call));
-            }
-            Rule::expression => {
-                let value = parse_expression(inner, &local_ctx)?;
-                return Ok(IfBodyResult::Value(value));
-            }
-            _ => {}
-        }
-    }
-
-    Err(ParseError::InternalError {
-        expected: "resource expression, module call, or value expression".to_string(),
-        context: "if body".to_string(),
-    })
-}
-
-/// Parse an if/else expression in value position (attribute values, not let bindings).
-///
-/// Unlike `parse_if_expr()` which returns `LetBindingRhs` (resources, module calls, or values),
-/// this function only returns `Value`. The condition must be a static Bool.
-/// An else clause is required when the condition is false (a value must always be determined).
-fn parse_if_value_expr(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-) -> Result<Value, ParseError> {
-    let mut inner = pair.into_inner();
-
-    // Parse the condition expression
-    let condition_pair = next_pair(&mut inner, "condition", "if value expression")?;
-    let condition_value = parse_expression(condition_pair, ctx)?;
-
-    // Ensure the condition is statically evaluable
-    if !is_static_value(&condition_value) {
-        return Err(ParseError::InvalidExpression {
-            line: 0,
-            message: "if condition depends on a runtime value; \
-                      condition must be statically known at parse time"
-                .to_string(),
-        });
-    }
-
-    let condition_value = evaluate_static_value(condition_value, ctx.config)?;
-
-    // Condition must be a Bool
-    let condition = match &condition_value {
-        Value::Bool(b) => *b,
-        other => {
-            return Err(ParseError::InvalidExpression {
-                line: 0,
-                message: format!("if condition must be a Bool value, got: {:?}", other),
-            });
-        }
-    };
-
-    // Parse the if body
-    let if_body_pair = next_pair(&mut inner, "if body", "if value expression")?;
-
-    // Check for else clause
-    let else_body_pair = inner.next();
-
-    if condition {
-        parse_if_body_value(if_body_pair, ctx)
-    } else if let Some(else_pair) = else_body_pair {
-        let else_body = first_inner(else_pair, "else body", "else clause")?;
-        parse_if_body_value(else_body, ctx)
-    } else {
-        Err(ParseError::InvalidExpression {
-            line: 0,
-            message: "if expression in value position requires an else clause \
-                      when condition is false"
-                .to_string(),
-        })
-    }
-}
-
-/// Parse the body of an if expression in value position and return only the value.
-fn parse_if_body_value(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-) -> Result<Value, ParseError> {
-    let mut local_ctx = ctx.clone();
-
-    for inner in pair.into_inner() {
-        match inner.as_rule() {
-            Rule::if_local_binding => {
-                let mut binding_inner = inner.into_inner();
-                let name = next_pair(&mut binding_inner, "binding name", "if local binding")?
-                    .as_str()
-                    .to_string();
-                let value = parse_expression(
-                    next_pair(&mut binding_inner, "binding value", "if local binding")?,
-                    &local_ctx,
-                )?;
-                local_ctx.set_variable(name, value);
-            }
-            Rule::expression => {
-                return parse_expression(inner, &local_ctx);
-            }
-            Rule::resource_expr | Rule::read_resource_expr | Rule::module_call => {
-                return Err(ParseError::InvalidExpression {
-                    line: 0,
-                    message: "resource expressions and module calls cannot be used \
-                              in if value expressions; use a let binding instead"
-                        .to_string(),
-                });
-            }
-            _ => {}
-        }
-    }
-
-    Err(ParseError::InternalError {
-        expected: "value expression".to_string(),
-        context: "if value expression body".to_string(),
-    })
 }
 
 fn parse_provider_block(
@@ -2967,34 +2109,6 @@ fn parse_resource_address(pair: pest::iterators::Pair<Rule>) -> Result<ResourceI
     let (provider, resource_type) = split_namespaced_id(&namespaced);
 
     Ok(ResourceId::with_provider(provider, resource_type, name))
-}
-
-/// Parse a string token into its literal value (without quotes).
-/// Only handles plain strings (no interpolation).
-fn parse_string_literal(pair: pest::iterators::Pair<Rule>) -> Result<String, ParseError> {
-    // string = single_quoted_string | double_quoted_string
-    let inner_pair = pair.into_inner().next().unwrap();
-
-    if inner_pair.as_rule() == Rule::single_quoted_string {
-        return Ok(inner_pair
-            .into_inner()
-            .next()
-            .map(|p| unescape_single_quoted(p.as_str()))
-            .unwrap_or_default());
-    }
-
-    // Double-quoted string
-    let mut result = String::new();
-    for part in inner_pair.into_inner() {
-        if part.as_rule() == Rule::string_part {
-            for inner in part.into_inner() {
-                if inner.as_rule() == Rule::string_literal {
-                    result.push_str(inner.as_str());
-                }
-            }
-        }
-    }
-    Ok(result)
 }
 
 /// Parse an import state block
@@ -4205,589 +3319,6 @@ fn parse_expression_eval(
 ) -> Result<EvalValue, ParseError> {
     let inner = first_inner(pair, "expression body", "expression")?;
     parse_coalesce_expr(inner, ctx)
-}
-
-fn parse_coalesce_expr(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-) -> Result<EvalValue, ParseError> {
-    let mut inner = pair.into_inner();
-    let first = next_pair(&mut inner, "pipe expression", "coalesce expression")?;
-    let value = parse_pipe_expr(first, ctx)?;
-
-    // If there's a ?? right-hand side, check if left is an unresolved reference
-    if let Some(rhs_pair) = inner.next() {
-        let default = parse_pipe_expr(rhs_pair, ctx)?;
-        match &value {
-            EvalValue::User(Value::ResourceRef { .. }) => Ok(default),
-            _ => Ok(value),
-        }
-    } else {
-        Ok(value)
-    }
-}
-
-fn parse_compose_expr(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-) -> Result<EvalValue, ParseError> {
-    let mut inner = pair.into_inner();
-    let first = next_pair(&mut inner, "primary expression", "compose expression")?;
-    let mut value = parse_primary_eval(first, ctx)?;
-
-    // Collect remaining primaries for >> composition
-    for rhs_pair in inner {
-        let rhs = parse_primary_eval(rhs_pair, ctx)?;
-
-        // Both sides must be Closures
-        if !value.is_closure() {
-            return Err(ParseError::InvalidExpression {
-                line: 0,
-                message: format!(
-                    "left side of >> must be a Closure (partially applied function), got {}",
-                    eval_type_name(&value)
-                ),
-            });
-        }
-        if !rhs.is_closure() {
-            return Err(ParseError::InvalidExpression {
-                line: 0,
-                message: format!(
-                    "right side of >> must be a Closure (partially applied function), got {}",
-                    eval_type_name(&rhs)
-                ),
-            });
-        }
-
-        // Build a composed closure: __compose__ with the chain stored in captured_args
-        // If the left side is already a __compose__, extend the chain; otherwise start a new one
-        let functions = if let EvalValue::Closure {
-            name,
-            captured_args,
-            ..
-        } = &value
-            && name == "__compose__"
-        {
-            let mut fns = captured_args.clone();
-            fns.push(rhs);
-            fns
-        } else {
-            vec![value, rhs]
-        };
-
-        value = EvalValue::closure("__compose__", functions, 1);
-    }
-
-    Ok(value)
-}
-
-fn parse_pipe_expr(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-) -> Result<EvalValue, ParseError> {
-    let mut inner = pair.into_inner();
-    let compose = next_pair(&mut inner, "compose expression", "pipe expression")?;
-    let mut value = parse_compose_expr(compose, ctx)?;
-
-    // Desugar pipe: `x |> f(args)` becomes `f(x, args)`
-    for func_call_pair in inner {
-        let mut fc_inner = func_call_pair.into_inner();
-        let func_name = next_pair(&mut fc_inner, "function name", "pipe function call")?
-            .as_str()
-            .to_string();
-        let extra_args: Result<Vec<Value>, ParseError> =
-            fc_inner.map(|arg| parse_expression(arg, ctx)).collect();
-        let extra_args = extra_args?;
-
-        // Check if the pipe target is a Closure variable. The pipe
-        // value (the running `value`) is appended last as an
-        // EvalValue, so a closure carried in the binding can finish
-        // applying through subsequent pipes.
-        if let Some(EvalValue::Closure {
-            name: fn_name,
-            captured_args,
-            remaining_arity,
-        }) = ctx.get_variable(&func_name)
-        {
-            let mut all_args: Vec<EvalValue> = extra_args
-                .iter()
-                .cloned()
-                .map(EvalValue::from_value)
-                .collect();
-            all_args.push(value.clone());
-            if all_args.iter().all(is_static_eval) {
-                value = crate::builtins::apply_closure_with_config(
-                    fn_name,
-                    captured_args,
-                    *remaining_arity,
-                    &all_args,
-                    ctx.config,
-                )
-                .map_err(|e| ParseError::InvalidExpression {
-                    line: 0,
-                    message: e,
-                })?;
-                continue;
-            }
-        }
-
-        // Build args for the non-closure dispatch path. The running
-        // pipe value must be a user-facing value here; a closure
-        // would mean the user piped a partial application into a
-        // non-closure-aware call, which we surface as a parse error.
-        let pipe_value = match value {
-            EvalValue::User(v) => v,
-            EvalValue::Closure {
-                ref name,
-                remaining_arity,
-                ..
-            } => {
-                return Err(ParseError::InvalidExpression {
-                    line: 0,
-                    message: format!(
-                        "cannot pipe a closure '{}' (still needs {} arg(s)) into '{}' \
-                         — finish the partial application first",
-                        name, remaining_arity, func_name
-                    ),
-                });
-            }
-        };
-        let mut args = extra_args;
-        args.push(pipe_value);
-
-        // Try to eagerly evaluate user-defined function calls
-        if ctx.user_functions.contains_key(&func_name) && args.iter().all(is_static_value) {
-            let user_fn = ctx.user_functions.get(&func_name).unwrap().clone();
-            value = EvalValue::from_value(evaluate_user_function(&user_fn, &args, ctx)?);
-        } else if let Some(arity) = crate::builtins::builtin_arity(&func_name) {
-            // Eagerly evaluate partial application for builtin pipe targets
-            if args.len() < arity && args.iter().all(is_static_value) {
-                let eval_args: Vec<EvalValue> =
-                    args.iter().cloned().map(EvalValue::from_value).collect();
-                value = crate::builtins::evaluate_builtin_with_config(
-                    &func_name, &eval_args, ctx.config,
-                )
-                .map_err(|e| ParseError::InvalidExpression {
-                    line: 0,
-                    message: format!("{}(): {}", func_name, e),
-                })?;
-            } else {
-                value = EvalValue::from_value(Value::FunctionCall {
-                    name: func_name,
-                    args,
-                });
-            }
-        } else {
-            value = EvalValue::from_value(Value::FunctionCall {
-                name: func_name,
-                args,
-            });
-        }
-    }
-
-    Ok(value)
-}
-
-fn parse_primary_eval(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-) -> Result<EvalValue, ParseError> {
-    // For primary, get inner content; otherwise process directly
-    let inner = if pair.as_rule() == Rule::primary {
-        first_inner(pair, "value", "primary expression")?
-    } else {
-        pair
-    };
-
-    match inner.as_rule() {
-        Rule::resource_expr => {
-            // Resource expressions cannot be used as attribute values (only valid in top-level let bindings)
-            Err(ParseError::InvalidExpression {
-                line: 0,
-                message: "Resource expressions can only be used in let bindings".to_string(),
-            })
-        }
-        Rule::list => {
-            let items: Result<Vec<Value>, ParseError> = inner
-                .into_inner()
-                .map(|item| parse_expression(item, ctx))
-                .collect();
-            Ok(EvalValue::from_value(Value::List(items?)))
-        }
-        Rule::map => {
-            let mut map: IndexMap<String, Value> = IndexMap::new();
-            let mut nested_blocks: IndexMap<String, Vec<Value>> = IndexMap::new();
-            for entry in inner.into_inner() {
-                match entry.as_rule() {
-                    Rule::map_entry => {
-                        let mut entry_inner = entry.into_inner();
-                        let key_pair = next_pair(&mut entry_inner, "map key", "map entry")?;
-                        let key = extract_key_string(key_pair)?;
-                        let value = parse_expression(
-                            next_pair(&mut entry_inner, "map value", "map entry")?,
-                            ctx,
-                        )?;
-                        map.insert(key, value);
-                    }
-                    Rule::nested_block => {
-                        let mut block_inner = entry.into_inner();
-                        let block_name =
-                            next_pair(&mut block_inner, "block name", "nested block in map")?
-                                .as_str()
-                                .to_string();
-                        let block_attrs = parse_block_contents(block_inner, ctx)?;
-                        nested_blocks
-                            .entry(block_name)
-                            .or_default()
-                            .push(Value::Map(block_attrs));
-                    }
-                    _ => {}
-                }
-            }
-            for (name, blocks) in nested_blocks {
-                map.insert(name, Value::List(blocks));
-            }
-            Ok(EvalValue::from_value(Value::Map(map)))
-        }
-        Rule::namespaced_id => {
-            // Namespaced identifier (e.g., aws.Region.ap_northeast_1)
-            // or resource reference (e.g., bucket.name)
-            // or arguments reference in module context (e.g., arguments.vpc_id)
-            let full_str = inner.as_str();
-            let parts: Vec<&str> = full_str.split('.').collect();
-
-            if parts.len() == 2 {
-                // Two-part identifier: could be resource reference or variable access
-                if ctx.get_variable(parts[0]).is_some() && !ctx.is_resource_binding(parts[0]) {
-                    // Variable exists but trying to access attribute on non-resource
-                    Err(ParseError::InvalidExpression {
-                        line: 0,
-                        message: format!(
-                            "'{}' is not a resource, cannot access attribute '{}'",
-                            parts[0], parts[1]
-                        ),
-                    })
-                } else if ctx.is_resource_binding(parts[0]) {
-                    // Known resource binding: treat as resource reference
-                    Ok(EvalValue::from_value(Value::resource_ref(
-                        parts[0].to_string(),
-                        parts[1].to_string(),
-                        vec![],
-                    )))
-                } else {
-                    // Unknown 2-part identifier: could be TypeName.value enum shorthand
-                    // Will be resolved during schema validation
-                    Ok(EvalValue::from_value(Value::String(format!(
-                        "{}.{}",
-                        parts[0], parts[1]
-                    ))))
-                }
-            } else if ctx.is_resource_binding(parts[0]) {
-                // 3+ part identifier where first part is a resource binding:
-                // chained field access (e.g., web.network.vpc_id)
-                Ok(EvalValue::from_value(Value::resource_ref(
-                    parts[0].to_string(),
-                    parts[1].to_string(),
-                    parts[2..].iter().map(|s| s.to_string()).collect(),
-                )))
-            } else {
-                // 3+ part identifier is a namespaced type (aws.Region.ap_northeast_1)
-                Ok(EvalValue::from_value(Value::String(full_str.to_string())))
-            }
-        }
-        Rule::boolean => {
-            let b = inner.as_str() == "true";
-            Ok(EvalValue::from_value(Value::Bool(b)))
-        }
-        Rule::float => {
-            let f: f64 = inner
-                .as_str()
-                .parse()
-                .map_err(|e| ParseError::InvalidExpression {
-                    line: inner.line_col().0,
-                    message: format!("invalid float literal: {e}"),
-                })?;
-            Ok(EvalValue::from_value(Value::Float(f)))
-        }
-        Rule::number => {
-            let n: i64 = inner
-                .as_str()
-                .parse()
-                .map_err(|e| ParseError::InvalidExpression {
-                    line: inner.line_col().0,
-                    message: format!("integer literal out of range: {e}"),
-                })?;
-            Ok(EvalValue::from_value(Value::Int(n)))
-        }
-        Rule::string => parse_string_value(inner, ctx).map(EvalValue::from_value),
-        Rule::function_call => {
-            let mut fc_inner = inner.into_inner();
-            let func_name = next_pair(&mut fc_inner, "function name", "function call")?
-                .as_str()
-                .to_string();
-            let args: Result<Vec<Value>, ParseError> =
-                fc_inner.map(|arg| parse_expression(arg, ctx)).collect();
-            let args = args?;
-
-            // Check if the name refers to a Closure variable (direct call on closure)
-            if let Some(EvalValue::Closure {
-                name: fn_name,
-                captured_args,
-                remaining_arity,
-            }) = ctx.get_variable(&func_name)
-                && args.iter().all(is_static_value)
-            {
-                let eval_args: Vec<EvalValue> =
-                    args.iter().cloned().map(EvalValue::from_value).collect();
-                return crate::builtins::apply_closure_with_config(
-                    fn_name,
-                    captured_args,
-                    *remaining_arity,
-                    &eval_args,
-                    ctx.config,
-                )
-                .map_err(|e| ParseError::InvalidExpression {
-                    line: 0,
-                    message: e,
-                });
-            }
-
-            // Try to eagerly evaluate user-defined function calls
-            if ctx.user_functions.contains_key(&func_name) && args.iter().all(is_static_value) {
-                let user_fn = ctx.user_functions.get(&func_name).unwrap().clone();
-                return evaluate_user_function(&user_fn, &args, ctx).map(EvalValue::from_value);
-            }
-
-            // Eagerly evaluate partial application (fewer args than arity → Closure)
-            if let Some(arity) = crate::builtins::builtin_arity(&func_name)
-                && args.len() < arity
-                && args.iter().all(is_static_value)
-            {
-                let eval_args: Vec<EvalValue> =
-                    args.iter().cloned().map(EvalValue::from_value).collect();
-                return crate::builtins::evaluate_builtin_with_config(
-                    &func_name, &eval_args, ctx.config,
-                )
-                .map_err(|e| ParseError::InvalidExpression {
-                    line: 0,
-                    message: format!("{}(): {}", func_name, e),
-                });
-            }
-
-            Ok(EvalValue::from_value(Value::FunctionCall {
-                name: func_name,
-                args,
-            }))
-        }
-        Rule::variable_ref => {
-            // variable_ref = { identifier ~ (field_access | index_access)* }
-            // field_access = { "." ~ identifier }
-            // index_access = { "[" ~ expression ~ "]" }
-            let mut parts = inner.into_inner();
-            let first_ident = next_pair(&mut parts, "identifier", "variable reference")?.as_str();
-
-            // Collect all access steps (field or index)
-            let access_steps: Vec<pest::iterators::Pair<Rule>> = parts.collect();
-
-            if access_steps.is_empty() {
-                // Simple variable reference (no access chain)
-                match ctx.get_variable(first_ident) {
-                    Some(val) => Ok(val.clone()),
-                    None => Ok(EvalValue::from_value(Value::String(
-                        first_ident.to_string(),
-                    ))),
-                }
-            } else {
-                // Build binding_name, attribute_name, and field_path from access steps.
-                // Index access (e.g., [0] or ["key"]) composes the binding name.
-                // Field access after the binding gives attribute_name and field_path.
-                let mut binding_name = first_ident.to_string();
-                let mut field_names: Vec<String> = Vec::new();
-                let mut in_field_phase = false;
-
-                for step in access_steps {
-                    match step.as_rule() {
-                        Rule::index_access => {
-                            if in_field_phase {
-                                // Index access after field access is not yet supported
-                                // (e.g., a.b[0] — would need runtime list indexing)
-                                return Err(ParseError::InvalidExpression {
-                                    line: 0,
-                                    message: "index access after field access is not supported"
-                                        .to_string(),
-                                });
-                            }
-                            // Parse the index expression
-                            let index_expr_pair =
-                                first_inner(step, "index expression", "index access")?;
-                            let index_value = parse_expression(index_expr_pair, ctx)?;
-                            // Compose the binding name: name[0] or name["key"]
-                            match &index_value {
-                                Value::Int(n) => {
-                                    binding_name = format!("{}[{}]", binding_name, n);
-                                }
-                                Value::String(s) => {
-                                    binding_name = format!("{}[\"{}\"]", binding_name, s);
-                                }
-                                other => {
-                                    return Err(ParseError::InvalidExpression {
-                                        line: 0,
-                                        message: format!(
-                                            "index access key must be an integer or string, got {:?}",
-                                            other
-                                        ),
-                                    });
-                                }
-                            }
-                        }
-                        Rule::field_access => {
-                            in_field_phase = true;
-                            let field_ident =
-                                first_inner(step, "field identifier", "field access")?;
-                            field_names.push(field_ident.as_str().to_string());
-                        }
-                        _ => {}
-                    }
-                }
-
-                if field_names.is_empty() {
-                    // Index access only, no field access (e.g., subnets[0])
-                    // Check if the composed binding name is a known variable
-                    match ctx.get_variable(&binding_name) {
-                        Some(val) => Ok(val.clone()),
-                        None => {
-                            // Return as ResourceRef with empty attribute_name
-                            // (will be resolved later)
-                            Ok(EvalValue::from_value(Value::resource_ref(
-                                binding_name,
-                                String::new(),
-                                vec![],
-                            )))
-                        }
-                    }
-                } else {
-                    let attribute_name = field_names.remove(0);
-                    Ok(EvalValue::from_value(Value::resource_ref(
-                        binding_name,
-                        attribute_name,
-                        field_names,
-                    )))
-                }
-            }
-        }
-        Rule::if_expr => parse_if_value_expr(inner, ctx).map(EvalValue::from_value),
-        Rule::expression => parse_expression_eval(inner, ctx),
-        _ => Ok(EvalValue::from_value(Value::String(
-            inner.as_str().to_string(),
-        ))),
-    }
-}
-
-fn parse_string_value(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-) -> Result<Value, ParseError> {
-    use crate::resource::InterpolationPart;
-
-    // string = single_quoted_string | double_quoted_string
-    let inner_pair = first_inner(pair, "string content", "string")?;
-
-    if inner_pair.as_rule() == Rule::single_quoted_string {
-        // Single-quoted: literal only, no interpolation
-        let content = inner_pair
-            .into_inner()
-            .next()
-            .map(|p| unescape_single_quoted(p.as_str()))
-            .unwrap_or_default();
-        return Ok(Value::String(content));
-    }
-
-    // Double-quoted string (original behavior)
-    let mut parts: Vec<InterpolationPart> = Vec::new();
-    let mut has_interpolation = false;
-
-    for part in inner_pair.into_inner() {
-        if part.as_rule() == Rule::string_part {
-            let inner = first_inner(part, "string content", "string_part")?;
-            match inner.as_rule() {
-                Rule::string_literal => {
-                    let s = unescape_string(inner.as_str());
-                    parts.push(InterpolationPart::Literal(s));
-                }
-                Rule::interpolation => {
-                    has_interpolation = true;
-                    let expr_pair =
-                        first_inner(inner, "interpolation expression", "interpolation")?;
-                    let value = parse_expression(expr_pair, ctx)?;
-                    parts.push(InterpolationPart::Expr(value));
-                }
-                _ => {}
-            }
-        }
-    }
-
-    if has_interpolation {
-        Ok(Value::Interpolation(parts))
-    } else {
-        // No interpolation — collapse to a plain String
-        let s = parts
-            .into_iter()
-            .map(|p| match p {
-                InterpolationPart::Literal(s) => s,
-                _ => unreachable!(),
-            })
-            .collect::<String>();
-        Ok(Value::String(s))
-    }
-}
-
-/// Handle escape sequences in string literals
-fn unescape_string(s: &str) -> String {
-    let mut result = String::with_capacity(s.len());
-    let mut chars = s.chars();
-    while let Some(c) = chars.next() {
-        if c == '\\' {
-            match chars.next() {
-                Some('n') => result.push('\n'),
-                Some('r') => result.push('\r'),
-                Some('t') => result.push('\t'),
-                Some('"') => result.push('"'),
-                Some('\\') => result.push('\\'),
-                Some('$') => result.push('$'),
-                Some(other) => {
-                    result.push('\\');
-                    result.push(other);
-                }
-                None => result.push('\\'),
-            }
-        } else {
-            result.push(c);
-        }
-    }
-    result
-}
-
-/// Handle escape sequences in single-quoted string literals.
-/// Only `\'` and `\\` are recognized as escape sequences.
-fn unescape_single_quoted(s: &str) -> String {
-    let mut result = String::with_capacity(s.len());
-    let mut chars = s.chars();
-    while let Some(c) = chars.next() {
-        if c == '\\' {
-            match chars.next() {
-                Some('\'') => result.push('\''),
-                Some('\\') => result.push('\\'),
-                Some(other) => {
-                    result.push('\\');
-                    result.push(other);
-                }
-                None => result.push('\\'),
-            }
-        } else {
-            result.push(c);
-        }
-    }
-    result
 }
 
 /// Build an `UndefinedIdentifier` error enriched with the best
